Force-update output bindings on first update.
Ordinarily this logic only ran on property changes, and since it is usually used to bind to a view property like `OuterSize` which updates late anyway, it was never obvious that there was a problem.

However, if a _rebind_ occurs such that the view is bound to a new context, then the new context may never get updated because the view property is no longer changing.

The logic seems correct to auto-update when initially bound, as this is what's done implicitly for input bindings as well.