Add a base type converter for conversions that are "too specific".
This will walk up the class hierarchy to find less-granular match, and is useful for some odd corner cases like binding to `{this}` on an `Item` or `StardewValley.Object` type, where there is no specific conversion for e.g. a `Pickaxe`.

We use dead-last priority for this to avoid any nasty surprises.