Allow closing menus by clicking outside.
To prevent this from going haywire in the case of floating elements, negative margins, transparent elements, empty layouts and anything else that's not perfectly vanilla, we need a definitive (and recursive) visibility test. Since this only needs to run on clicks, the change adds a new set of methods to deal specifically with _visible_ hit tests, favoring accuracy over speed.

The result should be as close as possible to "can you actually see anything other than whitespace here" as opposed to vague information about possible boundaries.