Make `Close` method public via the controller API.
The `ViewMenu` itself already has this, and it does exactly what we want from an API point of view: provide a programmatic way to do the same thing that (for example) pressing the ESC key would do.

Useful for "picker" style submenus that are supposed to close when an item is picked, or an OK button is pressed, etc.