Mark bubbling events as handled in standard views.
If the view has built-in click or drag handling then we generally don't want these to bubble up to some higher-level view in the UI, such as a layout view. In all the relevant cases, if a user wants to handle the _outcome_ of these operations, then there is another event for it (typically `Change`, `Select`, etc.).

It's important to do this within the core library because there is no way for a user to do it after the fact, making it difficult to write catch-all handlers that don't trigger accidentally when the event shouldn't have bubbled.