Allow general click handlers to preempt left/right-click handlers.
Since there is a defined order, it is usually not expected behavior if a caller sets `Handled` (or returns `true`, via framework) in a `Click` handler, yet the subsequent `left-click` still runs.

This is a potentially breaking change but should still be the right thing to do in every case.