Add cursor attachments.
These allow an additional sprite to be drawn next to the cursor, similar to how "grabbed items" or "cursor items" work in the base game, but allowing any arbitrary sprite, not just `Item` instances.

Currently, these will always suppress tooltips to avoid messy overlap, although they are implemented as separate properties in case that needs to change.

Access is through the menu controller API since it is an exclusively menu-level concern and can't be managed intuitively by individual views.