Skip new-row setup on final loop iteration of grid layout.
There was a focus-search regression introduced in #31 that caused navigation from below a grid, up into the grid, to always snap to the last item instead of the item that was actually above the cursor, and this was due to having an extra row position in the list (`secondaryStartPositions`), causing the focus search to fail to find an item on that row and default to the last item.

We simply need to skip that part of the loop, and do only the primary-orientation row layout for the final row.