Make drop-down list scrollable, and add max list height.
This implicitly prevents the expanded list from overflowing the screen, and can be further restricted in height for usability/aesthetics.

Also fixes what looks like a bug in the `ScrollContainer` that would cause it to _expand_ to the container height, even if the scrollable itself was content-sized and the content was _smaller_ than the container. In general a scrollable should only constrain the layout height, not increase it; the latter is what `Stretch` length is for.

(The second change may be a breaking change for some layouts but should be easy to fix by adding a `layout` to the `scrollable`).